backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.9.0
pillow==11.3.0
python-decouple==3.8
django-allauth==65.11.2
django-bootstrap5==25.2
